Edg/src/main/java/org/generation/Edg/repository/UsuarioRepository.java
package org.generation.Edg.repository;

import java.util.List;
import java.util.Optional;

import org.generation.Edg.model.Usuario;
import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.stereotype.Repository;

@Repository
public interface UsuarioRepository extends JpaRepository<Usuario, Long> {

	public Optional<Usuario> findByEmail(String email);

	public List<Usuario> findAllByNomeContainingIgnoreCase(String nome);

	public Usuario findByNome(String nome);

}
